% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_areacir.R
\name{fo_area.cir}
\alias{fo_area.cir}
\title{Circle area}
\usage{
fo_area.cir(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Center point on the X axis.}

\item{y1}{Center point on the Y axis.}

\item{x2}{Point on the circumference on the X axis.}

\item{y2}{Point on the circumference on the Y axis.}
}
\value{
Numeric vector
}
\description{
Function which calculates an area of a
circle.
}
\examples{
fo_area.cir(1,1,2,2)
fo_area.cir(345,324,577,234)
}
