% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_areasqr.R
\name{fo_area.sqr}
\alias{fo_area.sqr}
\title{Square area}
\usage{
fo_area.sqr(x1, y1, x2, y2, x3, y3, x4, y4)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{x2}{Value of the second point on the X axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{x3}{Value of the third point on the X axis.}

\item{y3}{Value of the third point on the Y axis.}

\item{x4}{Value of the fourth point on the X axis.}

\item{y3}{Value of the fourth point on the Y axis.}
}
\value{
Numerical vector
}
\description{
Function which calculates an area of a square
based on its apexes.
}
\examples{
fo_area.sqr(0,0,1/2*sqrt(1),1/2*sqrt(1),0,sqrt(1),-1/2*sqrt(1),1/2*sqrt(1))
fo_area.sqr(0,0,5,0,5,5,0,5))
}
