% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_arearho.R
\name{fo_area_rho}
\alias{fo_area_rho}
\title{Rhombus area}
\usage{
fo_area_rho(xs, ys, r1, r2, alpha)
}
\arguments{
\item{xs}{Center point value on the X axis.}

\item{ys}{Center point value on the Y axis.}

\item{r1}{Length from the center to the apex.}

\item{r2}{Length from the center to the apex.}

\item{alpha}{Angle of rotation.}
}
\value{
Numeric value
}
\description{
Function which calculates an area of rhombus.
}
\examples{
fo_area_rho(0,0,3,5,30)
}
