% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_canhex.R
\name{fo_can.hex}
\alias{fo_can.hex}
\title{Hexagon creation}
\usage{
fo_can.hex(x1, y1, x2, y2, x3, y3, x4, y4, x5, y5, x6, y6)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{x2}{Value of the second point on the X axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{x3}{Value of the third point on the X axis.}

\item{y3}{Value of the third point on the Y axis.}

\item{x4}{Value of the fourth point on the X axis.}

\item{y4}{Value of the fourth point on the Y axis.}

\item{x5}{Value of the fifth point on the X axis.}

\item{y5}{Value of the fifth point on the Y axis.}

\item{x6}{Value of the sixth point on the X axis.}

\item{y6}{Value of the sixth point on the Y axis.}
}
\value{
Logical value
}
\description{
Function which tests whether the figure is a hexagon
based on its apexes.
}
\examples{
fo_can.hex(0,0,1,0)

}
