% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_canrec.R
\name{fo_can.rec}
\alias{fo_can.rec}
\title{Rectangle creation}
\usage{
fo_can.rec(x1, y1, x2, y2, x3, y3, x4, y4)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{x2}{Value of the second point on the X axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{x3}{Value of the third point on the X axis.}

\item{y3}{Value of the third point on the Y axis.}

\item{x4}{Value of the fourth point on the X axis.}

\item{y3}{Value of the fourth point on the Y axis.}
}
\value{
Logical value
}
\description{
Function which tests whether the figure is a rectangle
based on its apexes.
}
\examples{
fo_can.rec(0,0,4,0,4,5,0,5)
fo_can.rec(-5,-5,5,-5,5,5,-5,5)
}
