% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_centerplot.R
\name{fo_centerplot.hex}
\alias{fo_centerplot.hex}
\title{Hexagon center plot}
\usage{
fo_centerplot.hex(xs, ys, r, alpha = 0)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{x2}{Value of the second point on the X axis.}

\item{x3}{Value of the third point on the X axis.}

\item{x4}{Value of the fourth point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{y3}{Value of the third point on the Y axis.}

\item{y4}{Value of the fourth point on the Y axis.}

\item{y5}{Value of the fifth point on the Y axis.}

\item{y6}{Value of the sixth point on the Y axis.}

\item{x5}{Value of the fifth point on the X axis.}

\item{x6}{Value of the sixth point on the X axis.}
}
\value{
Plot
}
\description{
Function which plots hexagon made from its center.
}
\examples{
fo_centerplot.hex(0,0,3,45)

}
