% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_circir.R
\name{fo_cir.cir}
\alias{fo_cir.cir}
\title{Circle circumference}
\usage{
fo_cir.cir(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Center point on the X axis.}

\item{y1}{Center point on the Y axis.}

\item{x2}{Point on the circumference on the X axis.}

\item{y2}{Point on the circumference on the Y axis.}
}
\value{
Numerical vector
}
\description{
Function which calculates a circumference of a
circle.
}
\examples{
fo_cir.cir(1,1,2,2)
fo_cir.cir(123,254,234,764)
}
