% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_diasqr.R
\name{fo_dia.sqr}
\alias{fo_dia.sqr}
\title{Square diamiter}
\usage{
fo_dia.sqr(x1, y1, x2, y2, x3, y3, x4, y4)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{x2}{Value of the second point on the X axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{x3}{Value of the third point on the X axis.}

\item{y3}{Value of the third point on the Y axis.}

\item{x4}{Value of the fourth point on the X axis.}

\item{y4}{Value of the fourth point on the Y axis.}
}
\value{
Numeric vector
}
\description{
Function which calculates a diamiter of a square
based on its apexes.
}
\examples{
fo_dia.sqr(0,0,1/2*sqrt(1),1/2*sqrt(1),0,sqrt(1),-1/2*sqrt(1),1/2*sqrt(1))
fo_dia.sqr(0,0,5,0,5,5,0,5)
}
