% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_perhex.R
\name{fo_per.hex}
\alias{fo_per.hex}
\title{Hexagon perimiter}
\usage{
fo_per.hex(xs, ys, r, alpha = 0)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{x2}{Value of the second point on the X axis.}

\item{x3}{Value of the third point on the X axis.}

\item{x4}{Value of the fourth point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{y3}{Value of the third point on the Y axis.}

\item{y4}{Value of the fourth point on the Y axis.}

\item{y5}{Value of the fifth point on the Y axis.}

\item{y6}{Value of the sixth point on the Y axis.}

\item{x5}{Value of the fifth point on the X axis.}

\item{x6}{Value of the sixth point on the X axis.}
}
\value{
Numeric value
}
\description{
Function which calculates perimiter of a hexagon.
}
\examples{
fo_per.hex(1,0,3, 45)

}
