% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_perrho.R
\name{fo_per.rho}
\alias{fo_per.rho}
\title{Rhombus peremiter}
\usage{
fo_per.rho(xs, ys, r1, r2, alpha)
}
\arguments{
\item{xs}{Center point value on the X axis.}

\item{ys}{Center point value on the Y axis.}

\item{r1}{Lenght from the center to the apex.}

\item{r2}{Lenght from the center to the apex.}

\item{alpha}{Angle of rotation.}
}
\value{
Numerical vector
}
\description{
Function which peremiter of a rhombus.
}
\examples{
fo_per.rho(0,0,1,2,30)
}
