% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_perhex.R
\name{fo_per_hex}
\alias{fo_per_hex}
\title{Hexagon perimeter}
\usage{
fo_per_hex(xs, ys, r, alpha = 0)
}
\arguments{
\item{xs}{Center point value on the X axis.}

\item{ys}{Center point value on the Y axis.}

\item{r}{Lenght from the center to the apex.}

\item{alpha}{Angle of rotation.}
}
\value{
Numeric value
}
\description{
Function which calculates a perimeter of a hexagon.
}
\examples{
fo_per_hex(1,0,3, 45)

}
