% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_perpar.R
\name{fo_per_par}
\alias{fo_per_par}
\title{Parallelogram perimeter}
\usage{
fo_per_par(xs, ys, r1, r2, alpha, beta)
}
\arguments{
\item{xs}{Left down apex on the X axis.}

\item{ys}{Left down apex on the Y axis.}

\item{r1}{Lenght of the basis.}

\item{r2}{Length of the side.}

\item{alpha}{Angle of rotation.}

\item{beta}{Angle between the basis and the left side of
the parallelogram.}
}
\value{
Numeric vector
}
\description{
Function which calculates a perimeter of a
parallelogram.
}
\examples{
fo_per_par(0,0,3,3,45,150)
}
