% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_pertri.R
\name{fo_per_tri}
\alias{fo_per_tri}
\title{Triangle perimeter}
\usage{
fo_per_tri(x1, y1, x2, y2, x3, y3)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{x2}{Value of the second point on the X axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{x3}{Value of the third point on the X axis.}

\item{y3}{Value of the third point on the Y axis.}
}
\value{
Numeric value
}
\description{
Function which calculates perimeter of a triangle.
}
\examples{
fo_per_tri(0,0,0,1,15,15)
}
