% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_plothex.R
\name{fo_plot.hex}
\alias{fo_plot.hex}
\title{Hexagon plot}
\usage{
fo_plot.hex(xs, ys, r, alpha = 0)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{x2}{Value of the second point on the X axis.}

\item{x3}{Value of the third point on the X axis.}

\item{x4}{Value of the fourth point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{y3}{Value of the third point on the Y axis.}

\item{y4}{Value of the fourth point on the Y axis.}

\item{y5}{Value of the fifth point on the Y axis.}

\item{y6}{Value of the sixth point on the Y axis.}

\item{x5}{Value of the fifth point on the X axis.}

\item{x6}{Value of the sixth point on the X axis.}
}
\value{
Plot
}
\description{
Function which plots a hexagon.
}
\examples{
fo_plot.hex(1,0,3, 90)

}
