% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_plotrho.R
\name{fo_plot.rho}
\alias{fo_plot.rho}
\title{Rhombus plot}
\usage{
fo_plot.rho(xs, ys, r1, r2, alpha)
}
\arguments{
\item{xs}{Center point value on the X axis.}

\item{ys}{Center point value on the Y axis.}

\item{r1}{Lenght from the center to the apex.}

\item{r2}{Lenght from the center to the apex.}

\item{alpha}{Angle of rotation.}
}
\value{
Plot
}
\description{
Function which plots a rhombus.
}
\examples{
fo_per.rho(0,0,3,5,30)
}
