% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_plottri.R
\name{fo_plot.tri}
\alias{fo_plot.tri}
\title{Triangle plot}
\usage{
fo_plot.tri(x1, y1, x2, y2, x3, y3)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{x2}{Value of the second point on the X axis.}

\item{y2}{Value of the second point on the Y axis.}

\item{x3}{Value of the third point on the X axis.}

\item{y3}{Value of the third point on the Y axis.}
}
\value{
Plot
}
\description{
Function which plots a triangle based on its apexes.
}
\examples{
fo_plot.tri(0, 0, 1, 0, 0, 1)
}
