% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_plothex.R
\name{fo_plot_hex}
\alias{fo_plot_hex}
\title{Hexagon plot}
\usage{
fo_plot_hex(xs, ys, r, alpha = 0)
}
\arguments{
\item{xs}{Center point value on the X axis.}

\item{ys}{Center point value on the Y axis.}

\item{r}{Lenght from the center to the apex.}

\item{alpha}{Angle of rotation.}
}
\value{
Plot
}
\description{
Function which plots a hexagon.
}
\examples{
fo_plot_hex(1,0,3, 90)

}
