% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_plotper.R
\name{fo_plot_par}
\alias{fo_plot_par}
\title{Parallelogram plot}
\usage{
fo_plot_par(xs, ys, r1, r2, alpha, beta)
}
\arguments{
\item{xs}{Left down apex on the X axis.}

\item{ys}{Left down apex on the Y axis.}

\item{r1}{Lenght of the basis.}

\item{r2}{Length of the side.}

\item{alpha}{Angle of rotation.}

\item{beta}{Angle between the basis and the left side of
the parallelogram.}
}
\value{
Plot
}
\description{
Function which plots a parallellogram.
}
\examples{
fo_plot_par(0,0,3,3,0,30)
}
