% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_radcir.R
\name{fo_rad.cir}
\alias{fo_rad.cir}
\title{Circle radius}
\usage{
fo_rad.cir(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Center point on the X axis.}

\item{y1}{Center point on the Y axis.}

\item{x2}{Point on the circumference on the X axis.}

\item{y2}{Point on the circumference on the Y axis.}
}
\value{
Numerical vector
}
\description{
Function which calculates a radius of a
circle.
}
\examples{
fo_rad.cir(1,1,2,2)
}
