% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthline.R
\name{length.line}
\alias{length.line}
\title{Length line}
\usage{
\method{length}{line}(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{x2}{Value of the second point on the X axis.}

\item{y2}{Value of the second point on the Y axis.}
}
\value{
Numeric value
}
\description{
Function which caculates length of a line
based on the position of its ends.
}
\examples{
length.line(0,-10,0,10)

}
