% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthline.R
\name{length_line}
\alias{length_line}
\title{Length line}
\usage{
length_line(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Value of the first point on the X axis.}

\item{y1}{Value of the first point on the Y axis.}

\item{x2}{Value of the second point on the X axis.}

\item{y2}{Value of the second point on the Y axis.}
}
\value{
Numeric value
}
\description{
Function which calculates length of a line
based on the position of its end points.
}
\examples{
length_line(0,-5,0,5)

}
