% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fo_areahex.R
\name{fo_area.hex}
\alias{fo_area.hex}
\title{Hexagon area}
\usage{
fo_area.hex(xs, ys, r, alpha = 0)
}
\arguments{
\item{x1}{Value of the center point on the X axis.}

\item{x2}{Value of the center point on the X axis.}

\item{x3}{Length of a diamiter.}
}
\value{
Numeric value
}
\description{
Function which calculates an area of a hexagon
based on its apexes.
}
\examples{
fo_area.hex(1,0,3, 45)

}
