% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_centerplot.R
\name{fo_centerplot.hex}
\alias{fo_centerplot.hex}
\title{Hexagon center plot}
\usage{
fo_centerplot.hex(xs, ys, r, alpha = 0)
}
\arguments{
\item{xs}{Center point value on the X axis.}

\item{ys}{Center point value on the Y axis.}

\item{r}{Lenght from the center to the apex.}

\item{alpha}{Angle of rotation.}
}
\value{
Plot
}
\description{
Function which plots hexagon made from its center.
}
\examples{
fo_centerplot.hex(0,0,3,45)

}
